backend/requirements.txt
fastapi
uvicorn
python-multipart
httpx
pydantic
python-dotenv
aiofiles
gitpython
google-generativeai